feat(typing): add type definition file
