+++ environment.d.ts
@@ -1,0 +1,11 @@
+import "./service/environment.d.ts";
+
+declare module "cron" {
+  interface CronTime {
+    toString: () => string;
+  }
+
+  interface CronJob {
+    cronTime: CronTime;
+  }
+}
