+++ tsconfig.build.json
@@ -4,5 +4,5 @@
   "compilerOptions": {
     "baseUrl": "."
   },
-  "include": ["./"]
+  "include": ["."]
 }
