+++ tsconfig.json
@@ -2,6 +2,7 @@
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "yggdrasil",
   "compilerOptions": {
+    "baseUrl": ".",
     "incremental": true,
     "target": "es2021",
     "lib": ["esnext"],
